Move hyperion zebra setup plans into phase1 module, add lint direction rules
