+++ src/mx_bluesky/phase1/__init__.py
@@ -1,0 +1,1 @@
+"""For plans used in Hyperion and i04 only"""
